Add files via upload
Hi everyone,

I'm reaching out to the community today because I'm having some trouble with my Android app. It has five sections, and the one on the right is dedicated to taking notes. Unfortunately, whenever I try to click the "Add note" button, the app crashes instantly. This is instead of opening a new page for composing a note, which is what it's supposed to do.

I'd be incredibly grateful if anyone could offer some assistance in troubleshooting this issue. Any insights or suggestions would be greatly appreciated!

Thank you in advance for your time and help!

Best regards,

Malek